restrictedarea/requirements.txt
pika
pytz
numpy
opencv-python
pillow
redis
matplotlib>=3.8.0